moved SsoAuthProvider start/finish functions into a utility module
